refactor(LinkedClass): move class linking functionality into Schedule component
